mod 10=9 forcing logging 16 shorts : 100Hz at 867ms, 10Hz at 286ms

+++ AeroQuad/AeroQuad.ino
@@ -1585,7 +1585,7 @@ void loop () {
 //          }
 //          LogValueSpace((float)batteryData[0].voltage/100.0);
           for (byte motor = 0; motor < 16; motor++) {
-            LogValueSpace(1L);
+            LogValueSpace((short)1);
           }
         
           LOG_SERIAL.println();
